requirements.txt
fastapi[all]
uvicorn
pyyaml
redis
pyjwt
pymysql
tortoise-orm[asyncmy]
zmail
httpx
loguru
werkzeug
dbutils
uvloop
